### 2024-01-20 #6254 @editorialbot
[PRE REVIEW]: QSOnic: fast quasar continuum fitting
**Submitting author:** @p-slash (<a href="http://orcid.org/0000-0001-7336-8912">Naim Göksel Karaçaylı</a>)
**Repository:** https://github.com/p-slash/qsonic
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.8.2
