### 2024-01-29 #6254 @p-slash
Hi @ivastar ,

Here are my recommendations:

César Ramírez-Pérez - [cramirez@ifae.es](mailto:cramirez@ifae.es)
Corentin Ravoux - [corentin.ravoux@clermont.in2p3.fr](mailto:corentin.ravoux@clermont.in2p3.fr)

They have not used qsonic, but they are familiar with the algorithm and its application to Lyman-alpha forest analysis through another package called picca.
