### 2024-02-15 d9d0723
grammar fixes and extra paragraph for picca
