+++ docs/source/paper.md
@@ -25,15 +25,17 @@ bibliography: paper.bib
 
 # Summary
 
-Distant quasars shows absorption features at wavelengths shorter than the Ly$\alpha$ emission line in their spectrum due to intervening neutral hydrogen gas. The Ly$\alpha$ forest technique uses these absorption lines to map out the matter distribution in vast volumes when the universe was 1.5--3 billion years old ($2\lesssim z \lesssim 5$). This map of the cosmos successfully constrained cosmological models using the Baryon Acoustic Oscillations feature in the 3D correlation function [@bourbouxCompletedSDSSIVExtended:2020], various dark matter models [@irsicConstraintsWDM:2017] and the sum of the neutrino masses [@palanqueDelabrouilleNeutrinoMass:2015] using the 1D Ly$\alpha$ forest power spectrum [@karacayliOptimal1DLy:2022; @karacayliOptimal1DLy:2024].
+Distant quasars show absorption features at wavelengths shorter than the Ly$\alpha$ emission line in their spectrum due to intervening neutral hydrogen gas. The Ly$\alpha$ forest technique uses these absorption lines to map out the matter distribution in vast volumes when the universe was 1.5--3 billion years old ($2\lesssim z \lesssim 5$). This map of the cosmos successfully constrained cosmological models using the Baryon Acoustic Oscillations feature in the 3D correlation function [@bourbouxCompletedSDSSIVExtended:2020], various dark matter models [@irsicConstraintsWDM:2017] and the sum of the neutrino masses [@palanqueDelabrouilleNeutrinoMass:2015] using the 1D Ly$\alpha$ forest power spectrum [@karacayliOptimal1DLy:2022; @karacayliOptimal1DLy:2024].
 
-The density of the intervening gas is infered from the transmitted flux fraction which is the absorption depth relative to the unabsorbed quasar continum. The field has devised a simple continuum model to homogenize the analysis across quasars so that the errors in estimated continuum are contained to well-known modes. This model assumes each quasar continuum $C_q$ is a scaling of a mean quasar continuum shape $\bar C$ by a polynomial of $\ln\lambda$. For example, an order 1 polynomial introduces amplitude $a_q$ and slope $b_q$ terms for each quasar such that $C_q = (a_q + b_q \ln \lambda) \bar{C}$. The quasar continuum fitting algorithm is then an iterative solution that minimizes $\chi^2$ for each quasar and updates the mean quasar continuum using the stack of all quasar continua after every iteration [@bourbouxCompletedSDSSIVExtended:2020; @karacayliOptimal1DLy:2024]. This methods requires efficient parallelization due to computational time and memory needed.
+The density of the intervening gas is inferred from the transmitted flux fraction which is the absorption depth relative to the unabsorbed quasar continuum. The field has devised a simple continuum model to homogenize the analysis across quasars so that the errors in the estimated continuum are contained to well-known modes. This model assumes each quasar continuum $C_q$ is a scaling of a mean quasar continuum shape $\bar C$ by a polynomial of $\ln\lambda$. For example, an order 1 polynomial introduces amplitude $a_q$ and slope $b_q$ terms for each quasar such that $C_q = (a_q + b_q \ln \lambda) \bar{C}$. The quasar continuum fitting algorithm is then an iterative solution that minimizes $\chi^2$ for each quasar and updates the mean quasar continuum using the stack of all quasar continua after every iteration [@bourbouxCompletedSDSSIVExtended:2020; @karacayliOptimal1DLy:2024]. This method requires efficient parallelization due to computational time and memory needed.
 
 
 # Statement of need
 
[…]
