### 2024-02-19 #6373 @editorialbot
[REVIEW]: QSOnic: fast quasar continuum fitting
**Submitting author:** @p-slash (<a href="http://orcid.org/0000-0001-7336-8912">Naim Göksel Karaçaylı</a>)
**Repository:** https://github.com/p-slash/qsonic
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.0
