### 2024-02-20 #6373 @corentinravoux
I have a COI with this review, I work in the same collaboration than the main author (DESI), and we are co-author of some papers (Karaçaylı et al. 2024, Ravoux et al. 2023). I am contributing and using the picca code mentioned in the paper.
However, I think I am able to provide an impartial assessment of this work, as I am open to using alternative continuum fitting algorithm, and I never used QSOnic. I request to waive this COI if this is possible.
