### 2024-02-20 #6373 @corentinravoux
Comments on the joss-paper (I used the download version):
- For the reference, I would add Armengaud et al. 2017 for the dark matter constraints, and Ravoux et al. 2023 for the measurements.
- The sentence "The field has devised ..." is unclear. Do you mean recent studies ? And to which "well-known modes" are you referring ?
- "a polynomial of ln 𝜆" Maybe I am wrong but i thought that for DESI, it is just "𝜆" ?
- "a delete-one Jackknife estimate for the covariance matrix used in pipeline noise calibration correction calculation" Here, do you mean the picca equivalent of eta ? This end of sentence might be too technical for the statement of need here.
- When you say that QSOnic is built on the same algorithm than picca, do you mean that you used part of the software, or that the principle of the algorithm is the same ? Please clarify that on the text.
