### 2024-02-20 p-slash/qsonic#92 @corentinravoux
JOSS Review
General comments regarding QSOnic, based on the review on JOSS subsmission. Complete review available in https://github.com/openjournals/joss-reviews/issues/6373
