### 2024-02-20 p-slash/qsonic#92 @corentinravoux
I have performed the first start of the review. The performance of QSOnic is excellent; there is almost a factor two in terms of reading and continuum fitting compared to picca.

I have tested the scripts qsonic-fit and qsonic-calib on DESI data with EDR data and (not yet public) Y1 data. I also tested the DLA and Sky line masking used in my analysis. I did not test computing on mocks, spectra reading, or the computation of P1D based on the outputs of QSOnic. Please point me to the tests you want me to perform so that I can give you a broader view of the package.

I have only one general comment regarding your software: It is unclear to me how you place QSOnic with respect to picca. Do you want to replace the delta_extraction part of picca ? My main problem is that if we maintain two software, all the picca improvements cannot be easily translated to QSOnic. Maybe state more clearly in your statement of need how you see the interaction between your software and picca in the future years.

Small comments:
- QSOnic is currently not compatible with python 3.11. Is there a specific reason for this? If not, please change the compatibility.
- In the documentation, the quick start command does not work as is; for example, the "--coadd-arms" option needs "before". Furthermore the argument "--no-interations" does not exist anymore.
- Please state the dependencies of your package inside the documentation, in the README.md, or both.
- Please provide more examples of script executions in the documentation, such as Raw mocks, DESI-EDR, and DESI-Y1. Please add a script if your software can run on eBOSS data.
- Provide a script of qsonic-calib, even if the arguments are close to qsonic-fit. Please extend the description in the documentation of what qsonic-calib and qsonic-fit are doing.
- Please provide guidelines for people who want to contribute, report issues, or seek help using your software. Add that inside the documentation, or the README.md, or both
- It would be valuable to transform part of your documentation into an available jupyter-notebook stored inside your package. I am particularly thinking about the «Reading spectra» and «Look into output files». It is way easier to test those functionnality if you provide data examples (it can be mocks) and notebooks.
