### 2024-02-20 #6373 @corentinravoux
The review process is quite faster than I expected. I summarized my comments in the following issue: https://github.com/p-slash/qsonic/issues/92.
There are some comments that are blocking the completion on my checklist. I let time to the main author of QSOnic to address them before I validate all points.
