### 2024-03-08 p-slash/qsonic#93 @gcalderone
JOSS review
Review for the JOSS submitted paper (original checklist: https://github.com/openjournals/joss-reviews/issues/6373);

The software aims to estimate a non-absorbed mean quasar continuum bluewards of the Lya, and claims to be faster than picca (although providing similar functionalities).

The software appears to be at a rather mature stage, still I believe it would benefit from a **major revision** before the paper is accepted, with changes involving mainly the documentation and the test suite.

## Major changes:

- The software can't be tested on user data unless the user is a member of the DESI collaboration and knows where/how to download the data.  To be more specfic:
  - how can I download the spectra in `desi-y5-1000/spectra-16/`?
  - what is/how can I download the v0 catalog for iron release?
  - what is an imaginary Y5 DESI quickquasars mock?

  Hence I suggest to either declare that the software is targeted exclusively to the members of the DESI collaboration, or to add a complete example involving publicly available data;

- The test suite fails with the following errors:
```
======================================================================== short test summary info =========================================================================
FAILED tests/test_calibration.py::test_calibrations - TypeError: NoiseCalibrator.__init__() got an unexpected keyword argument 'add_varlss'
FAILED tests/test_catalog.py::TestCatalog::test_read_validate_adjust_column_names - TypeError: _validate_adjust_column_names() got an unexpected keyword argument 'is_tile'
FAILED tests/test_catalog.py::TestCatalog::test_prime_catalog - TypeError: _prime_catalog() got an unexpected keyword argument 'is_tile'
FAILED tests/test_io.py::TestIOReading::test_read_spectra_onehealpix - AssertionError: assert 2 == 5
FAILED tests/test_mpi_utils.py::TestMPIUtils::test_balance_load - TypeError: balance_load() missing 1 required positional argument: 'mpi_rank'
FAILED tests/test_picca_continuum.py::TestVarLSSFitter::test_add - AttributeError: 'VarLSSFitter' object has no attribute 'calculate_subsampler_stats'. Did you mean: '_calc_subsampler_stats'?
FAILED tests/test_qsonic_fit.py::TestQsonicFit::test_parser - SystemExit: 2
FAILED tests/test_spectrum.py::TestSpectrum::test_simple_coadd - AttributeError: 'Spectrum' object has no attribute 'simple_coadd'
========================================================== 8 failed, 23 passed, 7 skipped, 2 warnings in 8.00s ===========================================================
```

- In the paper, the Author claims that QSOnic is *"faster by an order of magnitude"* w.r.t picca, but no evidence is presented to support the statement.  In particular:
  - Is the improvement mainly due to parallelization? or to a different algorithm?   In the former case, how does performance scale with number of processors?
  - Can a fair comparison be performed?  I.e. can the continuum parametrization of QSOnic (with two diversity parameters) be reduced to the picca one?
  - Are the obtained results compatible with picca?  or are they significantly different (making any performance comparison difficult to interpret) ?

  I would suggest to add a complete example using both QSOnic and picca to support the claim.  As an alternative, I suggest to adopt a milder statement such as "may be faster than picca due to the adopted parallelization".

## Minor changes

- The installation instructions seems to be targeted mainly to NERSC users.  I would suggest to move the *Install from PyPi* and *General* sections at the very beginning and to explicitly mention the main dependencies. E.g.
  - is python v3.10 the only accepted version? or is a later version supposed to work as well?
  - is MPI necessary?  If not, how to run the software without it?

- There probably is a typo in the mpi4py installation instructions: it should be `MPICC="mpicc -shared"` rather than `cc`.  If this is not the case it should be clearly stated;

- The "quick start" example shows an `srun` invocation.  Is slurm required to run the software?  If yes it should be clearly stated, otherwise the "quick start" example should not require a job scheduling system;

- In the "How it works" section of the documentation, what do *(1)* and *(2)* means in the first equation?  Are they the extreme of the wavelength range?

- The documentation does not corresponds to the software, e.g. the `get_spectra_reader_function` is not defined despite being mentioned.  Also, a few methods had their name changed , e.g. `read_spectra_onehealpix`;

- In the paper, the Author should clarify which criteria has been used to identify the parallelization domains (sky region / healpix?), as well as which part of the whole process is executed in parallel and which needs to be executed sequentially.
