### 2024-03-08 #6373 @gcalderone
I provided my comments in https://github.com/p-slash/qsonic/issues/93
