### 2024-03-13 ce29e36
move general instructions to top. explicitly write requirements.
