+++ docs/source/installation.rst
@@ -1,7 +1,36 @@
 Installation
 ============
 
-If you are on NERSC, I recommend using the pre-installed environment. This is installed on /global/common/software and has better performance at start up.
+General
+-------
+
+The following commands will create a conda environment and install qsonic.
+
+.. code-block:: shell
+    
+    conda create -n qsonic python=VERSION
+    conda activate qsonic
+    pip install qsonic
+
+- Change VERSION to the desired python version. The only restriction is ``VERSION>=3.7`` and there is no upper version restriction. However, I recommend using a python version that is one or two minor points behind the latest version to be safe.
+- Due to occasional compatibility problems between numpy, healpy and astropy, major dependencies are version limited to the following (see ``requirements.txt``):
+
+.. parsed-literal::
+
+    numpy~=1.24.0
+    numba~=0.57.0
+    scipy~=1.10.0
+    astropy~=5.2.0
+    healpy~=1.16.0
+    fitsio~=1.1.0
+    iminuit~=2.18.0
+    mpi4py 
+
+Note: MPI is necessary to run qsonic scripts.
+
+NERSC
+-----
+If you are a NERSC user, I recommend using the pre-installed environment. This is installed on /global/common/software and has better performance at start up.
 
 .. code:: shell
     
@@ -14,34 +43,22 @@ To have a Jupyter kernel, run the following command after activation:
     python -m ipykernel install --user --name qsonic --display-name qsonic
 
 
-Install from PyPi
------------------
-
-Explicit NERSC instructions for the lazy:
+*Explicit NERSC installation* instructions with pip:
 
 .. code-block:: shell
 
     module load cpu
     module load python
 
-    conda create -n qsonic python=3.10
+    conda create -n qsonic python
     conda activate qsonic
     MPICC="cc -shared" pip install --force-reinstall --no-cache-dir --no-binary=mpi4py mpi4py
     pip install qsonic
 
+Note: These instructions recompile ``mpi4py`` with NERSC specific `MPI libraries <https://docs.nersc.gov/development/languages/python/parallel-python/#mpi4py-in-your-custom-conda-environment>`_. ``cc`` command here is a NERSC compiler wrapper that provides optimal compiler options and paths to MPI libraries. Other platforms may use ``mpicc`` instead.
+
 If you want Jupyter kernel, you need to additionally install ``ipykernel`` into your environment and run the same command. See this `NERSC guide <https://docs.nersc.gov/services/jupyter/how-to-guides/>`_ on Jupyter.
 
-
-General
--------
-
-It is recommended that you create a separate conda environment and run
-
-.. code-block:: shell
-    
-    pip install qsonic
-
-You may also want to install required packages in ``requirements.txt`` while creating your conda environment, but pip should take care of it.
 
 Latest from the source
 ----------------------
@@ -59,9 +76,4 @@ For developers:
     git clone https://github.com/p-slash/qsonic.git
     cd qsonic
     pip install -e '.[dev]'
-
-NERSC
------
-
-``mpi4py`` needs special attention. Follow these `instructions <https://docs.nersc.gov/development/languages/python/parallel-python/#mpi4py-in-your-custom-conda-environment>`_ to clone a mpi4py installed conda environment on NERSC.
 
