### 2024-03-13 bab2ea8
clarify srun and mpirun
