### 2024-03-13 ac1a7a9
clarify (1) and (2) superscripts
