### 2024-03-13 1cd18a5
initial DESI EDR tutorial. Catalog is not correct!
