### 2024-03-14 efab234
EDR workaround
