### 2024-03-14 e40796f
update details on modules
