### 2024-03-14 746d97c
fixes to tutorial
