+++ docs/source/examples/simple_coadd_showcase.rst
@@ -38,7 +38,7 @@ reading all the quasar spectra in that file.
     split_catalog = np.split(catalog, s[1:])
 
     # Pick one healpix to illustrate
-    hpx_cat = split_catalog[0]
+    hpx_cat = split_catalog[1]
     healpix = hpx_cat['HPXPIXEL'][0]
 
     spectra_by_hpx = readerFunction(hpx_cat)
@@ -48,7 +48,7 @@ reading all the quasar spectra in that file.
 
 .. parsed-literal::
 
-    There are 207 spectra in healpix 0.
+    There are 71 spectra in healpix 9145.
 
 
 Let’s investigate one spectrum. Wavelength, flux and inverse variance
@@ -57,7 +57,7 @@ are stored as dictionaries similar to
 
 .. code:: python3
 
-    spec = spectra_by_hpx[0]
+    spec = spectra_by_hpx[3]
     print(spec.wave)
     print(spec.flux)
 
@@ -67,12 +67,12 @@ are stored as dictionaries similar to
     {'B': array([3600. , 3600.8, 3601.6, ..., 5798.4, 5799.2, 5800. ]),
      'R': array([5760. , 5760.8, 5761.6, ..., 7618.4, 7619.2, 7620. ]),
      'Z': array([7520. , 7520.8, 7521.6, ..., 9822.4, 9823.2, 9824. ])}
-    {'B': array([ 2.8399339 ,  5.5985265 ,  1.6996089 , ..., -0.02623173,
-            0.12586579,  0.17652267], dtype=float32),
-     'R': array([ 0.59253263, -4.5705295 ,  0.7594522 , ...,  2.133106  ,
-            0.19382767, -0.8677871 ], dtype=float32),
-     'Z': array([0.5383575 , 0.4969392 , 0.83893967, ..., 0.44942966, 0.642498  ,
-           0.68266016], dtype=float32)}
+    {'B': array([0.16013083, 2.1076498 , 6.495008  , ..., 2.2043223 , 1.6862453 ,
+        1.8163666 ], dtype=float32),
+     'R': array([-1.287594  ,  1.731283  ,  0.62619126, ...,  0.9037217 ,
+        1.3648763 ,  1.652868  ], dtype=float32),
+     'Z': array([0.83304965, 1.031328  , 1.6591258 , ..., 0.81355166, 1.0301682 ,
+        1.0923132 ], dtype=float32)}
 
 
 .. code:: python3
@@ -103,8 +103,8 @@ matter which arms are used to coadd.
 .. parsed-literal::
 
     {'brz': array([3600. , 3600.8, 3601.6, ..., 9822.4, 9823.2, 9824. ])}
-    {'brz': array([ 1.82258034,  4.42979435, -0.97420972, ...,  0.48654886,
-            0.54619423,  0.3381963 ])}
+    {'brz': array([0.16013082, 2.10764978, 6.49500805, ..., 0.81355166, 1.03016822,
+        1.0923132 ])}
 
 
 .. code:: python3
