### 2024-03-14 f5a37bf
updates to chi2 catalog based on edr
