+++ docs/source/examples/lookChi2Catalog.rst
@@ -1,6 +1,8 @@
 Look into output files
 ========================
 
+This tutorial builds on the DESI early data release example detailed in :ref:`Quick Start <edr example and workaround>`. I assume you have access to the EDR data, created ``QSO_cat_fuji_healpix_only_qso_targets_sv3_fix.fits`` file that fixes the compatibility issue, and obtained deltas.
+
 .. code:: python3
 
     import fitsio
@@ -9,7 +11,9 @@ Look into output files
 
 .. code:: python3
 
-    fchi2 = fitsio.FITS("Delta-co1/continuum_chi2_catalog.fits")[1]
+    # Point to output delta folder
+    output_delta_folder = "Delta-co1"
+    fchi2 = fitsio.FITS(f"{output_delta_folder}/continuum_chi2_catalog.fits")[1]
     fchi2
 
 .. parsed-literal::
@@ -19,14 +23,15 @@ Look into output files
       extension: 1
       type: BINARY_TBL
       extname: CHI2_CAT
-      rows: 471137
+      rows: 411359
       column info:
         TARGETID            i8  
         Z                   f4  
         HPXPIXEL            i8  
-        MPI_RANK            i4  
-        MEANSNR             f4  
+        ARMS                S5  
+        MEANSNR             f4  array[3]
         RSNR                f4  
+        MPI_RANK            i4  
         CONT_valid          b1  
         CONT_chi2           f4  
         CONT_dof            i4  
@@ -54,71 +59,62 @@ Look into output files
 .. image:: ../_static/chi2cat_hist.png
 
 
-Now let us investigate the ``attributes.fits`` file, which contain mean continuum, varlss-eta and stacked flux values for all iterations. Note this version lacks the final extensions, but they are the same as shown from another file.
+Now let us investigate the ``attributes.fits`` file, which contains the mean continuum in ``CONT-i`` extensions, stacked fluxes in observed frame in ``STACKED_FLUX-i``, in rest-frame in ``STACKED_FLUX_RF-i`` extensions, and varlss-eta values in ``VAR_FUNC-i`` extenstions for all iterations.
 
 .. code:: python3
 
-    fattr = fitsio.FITS("Delta-co1/attributes.fits")
+    fattr = fitsio.FITS(f"{output_delta_folder}/attributes.fits")
     fattr
-
-
 
 
 .. parsed-literal::
 
     
       file: Delta-co1/attributes.fits
-      mode: READONLY
       extnum hdutype         hduname[v]
       0      IMAGE_HDU       
       1      BINARY_TBL      CONT-1
-      2      BINARY_TBL      VAR_FUNC-1
-      3      BINARY_TBL      CONT-2
-      4      BINARY_TBL      VAR_FUNC-2
-      5      BINARY_TBL      CONT-3
-      6      BINARY_TBL      VAR_FUNC-3
-      7      BINARY_TBL      CONT-4
-      8      BINARY_TBL      VAR_FUNC-4
-      9      BINARY_TBL      CONT-5
-      10     BINARY_TBL      VAR_FUNC-5
-      11     BINARY_TBL      CONT-6
-      12     BINARY_TBL      VAR_FUNC-6
-      13     BINARY_TBL      CONT-7
-      14     BINARY_TBL      VAR_FUNC-7
-
[…]
