+++ docs/source/paper.bib
@@ -86,6 +86,26 @@
     eprint = {https://academic.oup.com/mnras/advance-article-pdf/doi/10.1093/mnras/stae171/56151699/stae171.pdf},
 }
 
+
+@ARTICLE{ravouxDesiLyaP1d:2023,
+       author = {{Ravoux}, Corentin and {Abdul Karim}, Marie Lynn and {Armengaud}, Eric and {Walther}, Michael and {Kara{\c{c}}ayl{\i}}, Naim G{\"o}ksel and {Martini}, Paul and {Guy}, Julien and {Aguilar}, Jessica Nicole and {Ahlen}, Steven and {Bailey}, Stephen and {Bautista}, Julian and {Beltran}, Sergio Felipe and {Brooks}, David and {Cabayol-Garcia}, Laura and {Chabanier}, Sol{\`e}ne and {Chaussidon}, Edmond and {Chaves-Montero}, Jon{\'a}s and {Dawson}, Kyle and {de la Cruz}, Rodrigo and {de la Macorra}, Axel and {Doel}, Peter and {Fanning}, Kevin and {Font-Ribera}, Andreu and {Forero-Romero}, Jaime and {Gontcho A Gontcho}, Satya and {Gonzalez-Morales}, Alma X. and {Gordon}, Calum and {Herrera-Alcantar}, Hiram K. and {Honscheid}, Klaus and {Ir{\v{s}}i{\v{c}}}, Vid and {Ishak}, Mustapha and {Kehoe}, Robert and {Kisner}, Theodore and {Kremin}, Anthony and {Landriau}, Martin and {Le Guillou}, Laurent and {Levi}, Michael and {Luki{\'c}}, Zarija and {Magneville}, Christophe and {Meisner}, Aaron and {Miquel}, Ramon and {Moustakas}, John and {Mueller}, Eva-Maria and {Mu{\~n}oz-Guti{\'e}rrez}, Andrea and {Napolitano}, Lucas and {Nie}, Jundan and {Niz}, Gustavo and {Palanque-Delabrouille}, Nathalie and {Percival}, Will and {P{\'e}rez-R{\`a}fols}, Ignasi and {Pieri}, Matthew and {Poppett}, Claire and {Prada}, Francisco and {Ram{\'\i}rez P{\'e}rez}, C{\'e}sar and {Rossi}, Graziano and {Sanchez}, Eusebio and {Schlegel}, David and {Schubnell}, Michael and {Seo}, Hee-Jong and {Sinigaglia}, Francesco and {Tan}, Ting and {Tarl{\'e}}, Gregory and {Wang}, Ben and {Weaver}, Benjamin and {Y{\`e}che}, Christophe and {Zhou}, Zhimin},
+        title = "{The Dark Energy Spectroscopic Instrument: one-dimensional power spectrum from first Ly {\ensuremath{\alpha}} forest samples with Fast Fourier Transform}",
+      journal = {\mnras},
+     keywords = {intergalactic medium, large-scale structure of Universe, cosmology: observations, Astrophysics - Cosmology and Nongalactic Astrophysics},
+         year = 2023,
+        month = dec,
+       volume = {526},
+       number = {4},
+        pages = {5118-5140},
+          doi = {10.1093/mnras/stad3008},
+archivePrefix = {arXiv},
+       eprint = {2306.06311},
+ primaryClass = {astro-ph.CO},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2023MNRAS.526.5118R},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
 @misc{picca,
[…]
