### 2024-03-14 036ca52
qsonic calib example script and minor fix
