### 2024-03-15 1f86ba9
move contributing to main repo. create examples/index.rst. mark notebooks as download
