+++ docs/source/examples/index.rst
@@ -1,0 +1,10 @@
+Examples
+========
+
+.. toctree::
+    :maxdepth: 2
+
+    lookChi2Catalog
+    simple_coadd_showcase
+    desi_member_examples
+    *
