### 2024-03-15 a5e49f3
move links around in quickstart
