### 2024-03-15 92014a8
fix reading function in howitworks
