+++ docs/source/howitworks.rst
@@ -51,7 +51,7 @@ We determine if the data needs to be blinded in :meth:`qsonic.spectrum.Spectrum.
 
 We read masks in :func:`qsonic.scripts.qsonic_fit.mpi_read_masks`. See :mod:`qsonic.masks` for :class:`SkyMask <qsonic.masks.SkyMask>`, :class:`BALMask <qsonic.masks.BALMask>` and :class:`DLAMask <qsonic.masks.DLAMask>` file conventions.
 
-We read spectra in :func:`qsonic.scripts.qsonic_fit.mpi_read_spectra_local_queue`. This is done by reading all **quasar** spectra in a healpix file using the function :func:`qsonic.io.read_spectra_onehealpix`, which uses slices to read **only** the quasar spectra in a given file. Then, the analysis region is enforced by the given wavelength arguments, which also populate ``forest*`` attributes of :class:`qsonic.spectrum.Spectrum`. Even though underlying API can keep the entire spectrum, we drop all regions that are outside the analysis region to save memory. Furthermore, a right-side SNR (RSNR) cut is applied to remove quasars if such an option is passed while reading, but techinally this is part of priming.
+We read spectra in :func:`qsonic.scripts.qsonic_fit.mpi_read_spectra_local_queue`. This is done by reading all **quasar** spectra in a healpix file using the reader function returned by :func:`qsonic.io.get_spectra_reader_function`, which uses slices to read **only** the quasar spectra in a given file. Then, the analysis region is enforced by the given wavelength arguments, which also populate ``forest*`` attributes of :class:`qsonic.spectrum.Spectrum`. Even though underlying API can keep the entire spectrum, we drop all regions that are outside the analysis region to save memory. Furthermore, a right-side SNR (RSNR) cut is applied to remove quasars if such an option is passed while reading, but techinally this is part of priming.
 
 Priming
 -------
