### 2024-03-15 p-slash/qsonic#92 @p-slash
Thank you for the input. I made the changes requested. You can access the updated documentation here: https://qsonic.readthedocs.io/en/paper/index.html

> I have tested the scripts qsonic-fit and qsonic-calib on DESI data with EDR data and (not yet public) Y1 data. I also tested the DLA and Sky line masking used in my analysis. I did not test computing on mocks, spectra reading, or the computation of P1D based on the outputs of QSOnic.

That is plenty and great. Thanks!

> I have only one general comment regarding your software: It is unclear to me how you place QSOnic with respect to picca. Do you want to replace the delta_extraction part of picca ? My main problem is that if we maintain two software, all the picca improvements cannot be easily translated to QSOnic. Maybe state more clearly in your statement of need how you see the interaction between your software and picca in the future years.

I added this sentence towards the end of the paper: `DESI Ly$\alpha$ forest working group and its ancillary topical groups are responsible for coordinating the efforts between different packages.`

> Small comments:
>
>     * QSOnic is currently not compatible with python 3.11. Is there a specific reason for this? If not, please change the compatibility.

I relaxed the python version with a best-practice warning in https://qsonic.readthedocs.io/en/paper/installation.html .

>     * In the documentation, the quick start command does not work as is; for example, the "--coadd-arms" option needs "before". Furthermore the argument "--no-interations" does not exist anymore.

Fixed.

>     * Please state the dependencies of your package inside the documentation, in the README.md, or both.

Stated in installation.

>     * Please provide more examples of script executions in the documentation, such as Raw mocks, DESI-EDR, and DESI-Y1. Please add a script if your software can run on eBOSS data.

Now provided here: https://qsonic.readthedocs.io/en/paper/examples/desi_member_examples.html. Unfortunately, qsonic cannot run on eBOSS. I also added a tutorial for publicly available EDR data (https://qsonic.readthedocs.io/en/paper/quick_start.html#example-desi-early-data-release). Most of the qsonic development has been done after EDR, which introduces some compatibility issues. I provided a workaround that is needed for EDR analysis in the documentation.

>     * Provide a script of qsonic-calib, even if the arguments are close to qsonic-fit. Please extend the description in the documentation of what qsonic-calib and qsonic-fit are doing.

Extended on the main page. qsonic-calib example added to quick start.

>     * Please provide guidelines for people who want to contribute, report issues, or seek help using your software. Add that inside the documentation, or the README.md, or both

Added CONTRIBUTING.rst to the main repo page, and to the documentation.

>     * It would be valuable to transform part of your documentation into an available jupyter-notebook stored inside your package. I am particularly thinking about the «Reading spectra» and «Look into output files». It is way easier to test those functionnality if you provide data examples (it can be mocks) and notebooks.

Jupyter notebooks now can be downloaded from the documentation or from docs/nb.
