### 2024-03-15 #6373 @p-slash
Thanks @corentinravoux for the feedback. This is the second part of my response.

>     * For the reference, I would add Armengaud et al. 2017 for the dark matter constraints, and Ravoux et al. 2023 for the measurements.

Added.

>     * The sentence "The field has devised ..." is unclear. Do you mean recent studies ? And to which "well-known modes" are you referring ?

Paper now says: ` Recent studies have devised a simple continuum model to homogenize the analysis across quasars so that the errors in the estimated continuum are contained to well-known modes, such as the amplitude and slope of the quasar continuum.`

>     * "a polynomial of ln 𝜆" Maybe I am wrong but i thought that for DESI, it is just "𝜆" ?

We are still fitting for ln lambda for DESI. There was a version of picca that performed the continuum fitting on lambda, but that was eventually removed before EDR analyses took off.

>     * "a delete-one Jackknife estimate for the covariance matrix used in pipeline noise calibration correction calculation" Here, do you mean the picca equivalent of eta ? This end of sentence might be too technical for the statement of need here.

Paper now says: `It saves detailed intermediate data products including best-fit parameters for each quasar and the covariance matrix used in pipeline noise calibration correction ($\eta$ parameter in references [@bourbouxCompletedSDSSIVExtended:2020; @karacayliOptimal1DLy:2024]).`

>     * When you say that QSOnic is built on the same algorithm than picca, do you mean that you used part of the software, or that the principle of the algorithm is the same ? Please clarify that on the text.

Paper now says: `QSOnic` is built on the same principle of [picca](https://github.com/igmhub/picca) continuum fitting algorithm. Unlike `picca`, `QSOnic` has robust convergence criteria, provides a generic API for DESI quasar spectra, and saves more intermediate data products. Due to optimized IO routines, MPI parallelization and efficient numerical algorithms, it may be faster than `picca`.
