### 2024-03-18 36a8189
read program column in the catalog as optional
