### 2024-03-20 8e45b6c
save some columns from catalog in delta headers
