### 2024-03-20 30b7fab
reading uncoadded spectra
