+++ tests/test_io.py
@@ -87,6 +87,22 @@ class TestIOReading(object):
                 npt.assert_allclose(spec.flux[arm], data['flux'][arm][jj])
                 npt.assert_allclose(spec.ivar[arm], data['ivar'][arm][jj])
 
+    def test_read_spectra_onehealpix_spectra(self, my_setup_fits_spectra):
+        cat_by_survey, input_dir, xarms, data = my_setup_fits_spectra
+        _, w = np.unique(cat_by_survey['TARGETID'], return_index=True)
+
+        slist = qsonic.io.read_onehealpix_file_data_spectra(
+            cat_by_survey[w][['TARGETID', 'Z', 'HPXPIXEL', 'SURVEY']],
+            input_dir, xarms, True)
+
+        assert (len(slist) == cat_by_survey.size)
+        for jj, spec in enumerate(slist):
+            assert (spec.targetid == cat_by_survey['TARGETID'][jj])
+            for arm in xarms:
+                npt.assert_allclose(spec.wave[arm], data['wave'][arm])
+                npt.assert_allclose(spec.flux[arm], data['flux'][arm][jj])
+                npt.assert_allclose(spec.ivar[arm], data['ivar'][arm][jj])
+
 
 @pytest.fixture()
 def my_setup_fits(tmp_path, setup_data):
@@ -104,6 +120,49 @@ def my_setup_fits(tmp_path, setup_data):
     d = d / f"{pixnum}"
     d.mkdir()
     fname = d / "coadd-main-dark-8258.fits"
+
+    with fitsio.FITS(fname, 'rw', clobber=True) as fts:
+        fts.write(cat_by_survey, extname="FIBERMAP")
+        for arm in xarms:
+            shape = data['flux'][arm].shape
+            fts.write(data['wave'][arm], extname=f"{arm}_WAVELENGTH")
+            fts.write(data['flux'][arm], extname=f"{arm}_FLUX")
+            fts.write(data['ivar'][arm], extname=f"{arm}_IVAR")
+            fts.write(np.zeros(shape, dtype='i4'), extname=f"{arm}_MASK")
+
+    # return sorted data truth
+    # Sort the generated catalog first.
+    sort_idx = np.argsort(cat_by_survey, order="TARGETID")
+    cat_by_survey = cat_by_survey[sort_idx]
+    for key in ['flux', 'ivar', 'mask']:
+        for arm in xarms:
+            data[key][arm] = data[key][arm][sort_idx]
+
+    yield cat_by_survey, tmp_path, xarms, data
+
+    shutil.rmtree(tmp_path / "main")
+
+
+@pytest.fixture()
+def my_setup_fits_spectra(tmp_path, setup_data):
+    cat_by_survey, _, data = setup_data(5)
+    pixnum = 8259
+    cat_by_survey['HPXPIXEL'] = pixnum
+    cat_by_survey['TARGETID'][:3] = cat_by_survey['TARGETID'][0]
+    cat_by_survey['TARGETID'][3:] = cat_by_survey['TARGETID'][3]
+    cat_by_survey['EXPID'][:3] = np.arange(3)
+    cat_by_survey['EXPID'][3:] = np.arange(2)
+    xarms = data['wave'].keys()
+
+    d = tmp_path / "main"
+    d.mkdir()
+    d = d / "dark"
+    d.mkdir()
+    d = d / f"{pixnum//100}"
+    d.mkdir()
+    d = d / f"{pixnum}"
+    d.mkdir()
+    fname = d / "spectra-main-dark-8259.fits"
 
     with fitsio.FITS(fname, 'rw', clobber=True) as fts:
         fts.write(cat_by_survey, extname="FIBERMAP")
