+++ py/qsonic/io.py
@@ -7,6 +7,7 @@ import warnings
 
 import fitsio
 import numpy as np
+from numpy.lib.recfunctions import drop_fields, merge_arrays
 
 from qsonic import QsonicException
 import qsonic.spectrum
@@ -60,6 +61,8 @@ def add_io_parser(parser=None):
         "--coadd-arms", default="before",
         choices=["before", "after", "disable"],
         help="Coadds arms before or after continuum fitting or not at all.")
+    outgroup.add_argument(
+        "--save-exposures", help="Reads and saves exposures.")
     outgroup.add_argument(
         "--save-by-hpx", action="store_true",
         help="Save by healpix. If not, saves by MPI rank.")
@@ -560,6 +563,141 @@ def read_onetile_coaddfile_data(
         spectra_list.extend(
             qsonic.spectrum.generate_spectra_list_from_data(
                 cat_by_petal, data)
+        )
+
+    return spectra_list
+
+
+def _read_onehealpix_file_spectra(
+        targetids_by_survey, fspec, arms_to_keep, skip_resomat
+):
+    """Common function to read all exposures from a single spectra- FITS file.
+
+    Arguments
+    ---------
+    targetids_by_survey: :external+numpy:py:class:`ndarray <numpy.ndarray>`
+        Targetids_by_survey (used to be catalog). If data, split by survey and
+        contains only one survey.
+    fspec: str
+        Filename to open.
+    arms_to_keep: list(str)
+        Must only contain B, R and Z.
+    skip_resomat: bool
+        If true, do not read resomat.
+
+    Returns
+    ---------
+    data: dict( :external+numpy:py:class:`ndarray <numpy.ndarray>` )
+        Only quasar spectra are read into keywords wave, flux etc. Resolution
+        is read if present.
+    idx_cat: :external+numpy:py:class:`ndarray <numpy.ndarray>`
+        Indices in `targetids_by_survey` there were succesfully read.
+    fbrmap: :external+numpy:py:class:`ndarray <numpy.ndarray>`
+        Catalog with ``TARGETID, PETAL_LOC, FIBER, NIGHT, EXPID, TILEID``
+        columns. Note the size will not be equal to targetids_by_survey.size.
+
+    Raises
+    ---------
+    RuntimeWarning
+        If number of quasars in the healpix file does not match the catalog.
+    """
+    # Assume it is sorted
+    # cat_by_survey.sort(order='TARGETID')
+    fitsfile = fitsio.FITS(fspec)
+
+    fbrmap = fitsfile['FIBERMAP'].read(
+        columns=['TARGETID', 'PETAL_LOC', 'FIBER', 'NIGHT', 'EXPID', 'TILEID'])
+
+    common_targetids, _, idx_cat = np.intersect1d(
+        fbrmap['TARGETID'], targetids_by_survey, return_indices=True)
+    if (common_targetids.size != targetids_by_survey.size):
+        warnings.warn(
+            f"Error reading {fspec}. "
+            "Number of quasars in healpix does not match the catalog "
+            f"catalog:{targetids_by_survey.size} vs "
+            f"healpix:{common_targetids.size}!", RuntimeWarning)
+
+    idx_fbr = np.nonzero(np.isin(fbrmap['TARGETID'], targetids_by_survey))[0]
+    targetids_fbr = fbrmap['TARGETID'][idx_fbr]
+    idx_fbr = idx_fbr[targetids_fbr.argsort()]
+    data = {
+        'wave': {},
+        'flux': {},
+        'ivar': {},
+        'mask': {},
[…]
