### 2024-03-21 c19b177
save exposure implemented to qsonic-fit script
