+++ py/qsonic/scripts/qsonic_fit.py
@@ -74,7 +74,9 @@ def args_logic_fnc_qsonic_fit(args):
         (args.mock_analysis and args.tile_format,
             "Mock analysis in tile format is not supported."),
         (args.tile_format and args.save_by_hpx,
-            "Cannot save deltas in healpixes in tile format.")
+            "Cannot save deltas in healpixes in tile format."),
+        (args.save_exposures and args.tile_format,
+            "Cannot save exposures in tile format.")
     ]
 
     for c, msg in condition_msg:
@@ -292,6 +294,88 @@ def mpi_continuum_fitting(spectra_list, args, comm, mpi_rank):
     return spectra_list
 
 
+def mpi_read_exposures(spectra_list, args, maskers, comm, mpi_rank):
+    """Creates a local catalog from spectra_list and reads exposures. Coadding
+    of arms is always done with IVAR only as weights (exposures are not coadded
+    ). RSNR cut is still applied. CONT is copied from the exposure coadded
+    spectra.
+
+    Arguments
+    ---------
+    spectra_list: list(Spectrum)
+        Spectrum objects for the local MPI rank.
+    args: argparse.Namespace
+        Options passed to script.
+    maskers: list(Masks)
+        Mask objects from `qsonic.masks`.
+    comm: MPI.COMM_WORLD
+        Communication object for reducing data.
+    mpi_rank: int
+        Rank of the MPI process
+
+    Returns
+    ---------
+    exposure_spectra_list: list(Spectrum)
+        Spectrum objects for the local MPI rank. Note there will be duplicate
+        TARGETIDs.
+    """
+    # Initialize continuum fitter & global functions
+    logging.info("Reading exposures and applying RSNR cut.")
+    start_time = time.time()
+
+    local_catalog = np.hstack([_.catrow for _ in spectra_list])
+    idx_sort = local_catalog.argsort(order=['HPXPIXEL', 'TARGETID'])
+    assert all(np.arange(idx_sort.size) == idx_sort)
+    del idx_sort
+
+    # exposures cannot be saved in tile format
+    unique_pix, s = np.unique(local_catalog['HPXPIXEL'], return_index=True)
+    local_queue = np.split(local_catalog, s[1:])
+    del s, unique_pix
+
+    exposure_spectra_list = []
+    # Each process reads its own list
+    for cat in local_queue:
+        local_specs = qsonic.io.read_onehealpix_file_data_spectra(
+            cat, args.input_dir, args.arms, args.skip_resomat
+        )
+
+        for spec in local_specs:
+            spec.set_forest_region(
+                args.wave1, args.wave2, args.forest_w1, args.forest_w2)
+            spec.remove_nonforest_pixels()
+
+        exposure_spectra_list.extend([
+            spec for spec in local_specs
+            if spec.forestwave and spec.rsnr >= args.min_rsnr
+        ])
+
+    del local_queue
+
+    if args.coadd_arms != "disable":
+        logging.info("Coadding arms with pure IVAR weights.")
+        for spec in exposure_spectra_list:
+            spec.coadd_arms_forest()
+
+    # Noise and flux calibration
+    mpi_noise_flux_calibrate(exposure_spectra_list, args, comm, mpi_rank)
+
+    # Mask out
+    apply_masks(maskers, exposure_spectra_list, mpi_rank)
+
+    # Match & assign cont, forest weight
+    for spec in exposure_spectra_list:
+        i = np.nonzero(local_catalog['TARGETID'] == spec.targetid)[0]
+        spec.cont_params['cont'] = spectra_list[i].cont_params['cont']
+
+    etime = (time.time() - start_time) / 60  # min
[…]
