+++ docs/source/howitworks.rst
@@ -100,6 +100,12 @@ We check for short spectra once more, which is important if ``--coadd_arms disab
 
 We finally save delta files in BinaryTable format. Delta files are organized in MPI ranks. It is also possible to save by the original healpix numbers in the catalog, but this creates a lot of files.
 
+
+Additional notes
+----------------
+**Saving exposures**: This option can be enabled by passing ``--save-exposures``. This will read exposures after continuum fitting is done. Coadding of arms (if ``--coadd_arms != disable``) is always done with only IVAR as weights (exposures are not coadded). RSNR cut is still applied. Continuum is copied from the exposure coadded spectra. The extension name in the delta files will be ``TARGETID_ARM_EXPID``.
+
+
 .. [Bautista2017] Bautista J. E., et al., 2017, `A&A, 603, A12 <https://ui.adsabs.harvard.edu/abs/2017A%26A...603A..12B/abstract>`_
 .. [duMasdesBourboux2019] du Mas des Bourboux H., et al., 2019, `ApJ, 878, 47 <https://ui.adsabs.harvard.edu/abs/2019ApJ...878...47D>`_
 .. [duMasdesBourboux2020] du Mas des Bourboux H., et al., 2020, `ApJ, 901, 153 <https://ui.adsabs.harvard.edu/abs/2020ApJ...901..153D/abstract>`_
