### 2024-03-21 bee28dd
extend _read_onehealpix_file to read uncoadded to spectra.
