+++ py/qsonic/io.py
@@ -77,10 +77,10 @@ def get_spectra_reader_function(
 ):
     """ Returns a callable object (function) that returns a list of Spectrum
     objects for a given catalog of a single healpix. Essentially, a wrapper
-    for :meth:`read_onehealpix_file_mock` or :meth:`read_onehealpix_file_data`
-    functions. If for data 'SURVEY' column must be present and sorted when
-    calling the function this returns in the healpix grouping. Tile grouping
-    requires TILEID and PETAL_LOC.
+    for :meth:`read_onehealpix_file_mock` or
+    :meth:`read_onehealpix_file_data_coadd` functions. If for data 'SURVEY'
+    column must be present and sorted when calling the function this returns in
+    the healpix grouping. Tile grouping requires TILEID and PETAL_LOC.
 
     Arguments
     ---------
@@ -125,7 +125,7 @@ def get_spectra_reader_function(
 
     else:
         return functools.partial(
-            read_onehealpix_file_data,
+            read_onehealpix_file_data_coadd,
             input_dir=input_dir, arms_to_keep=arms_to_keep,
             skip_resomat=skip_resomat, program=program
         )
@@ -334,10 +334,10 @@ def _read_true_continuum(targetids, fspec):
 
 
 def _read_onehealpix_file(
-        targetids_by_survey, fspec, arms_to_keep, skip_resomat
+        targetids_by_survey, fspec, arms_to_keep, skip_resomat,
+        fbrmap_columns=['TARGETID']
 ):
-    """Common function to read a single FITS file for fibermaps with unique
-    TARGETIDs.
+    """Common function to read a single FITS file in DESI healpix grouping.
 
     Arguments
     ---------
@@ -350,6 +350,10 @@ def _read_onehealpix_file(
         Must only contain B, R and Z.
     skip_resomat: bool
         If true, do not read resomat.
+    fbrmap_columns: list(str), default: ['TARGETID']
+        Columns to read from FIBERMAP. For coadded files, it should be
+        ``['TARGETID']``. For uncoadded files (spectra-), you should read at
+        least ``[TARGETID, PETAL_LOC, FIBER, NIGHT, EXPID, TILEID]``.
 
     Returns
     ---------
@@ -358,25 +362,31 @@ def _read_onehealpix_file(
         is read if present.
     idx_cat: :external+numpy:py:class:`ndarray <numpy.ndarray>`
         Indices in `targetids_by_survey` there were succesfully read.
+    fbrmap: :external+numpy:py:class:`ndarray <numpy.ndarray>`
+        Catalog with ``columns`` from the FIBERMAP. Note the size will not be
+        equal to ``targetids_by_survey.size`` when reading exposures.
 
     Raises
     ---------
     RuntimeWarning
         If number of quasars in the healpix file does not match the catalog.
     """
-    # Assume it is sorted
-    # cat_by_survey.sort(order='TARGETID')
     fitsfile = fitsio.FITS(fspec)
 
-    fbrmap = fitsfile['FIBERMAP'].read(columns='TARGETID')
-    common_targetids, idx_fbr, idx_cat = np.intersect1d(
-        fbrmap, targetids_by_survey, assume_unique=True, return_indices=True)
+    fbrmap = fitsfile['FIBERMAP'].read(columns=fbrmap_columns)
+
+    common_targetids, _, idx_cat = np.intersect1d(
+        fbrmap['TARGETID'], targetids_by_survey, return_indices=True)
[…]
