### 2024-03-25 p-slash/qsonic#93 @gcalderone
Thank you!
You addressed all points, and I recommend acceptance after the following **minor changes** are applied:
- In the "Quick start" section you correctly mention the size of the dataset (80 TB).  Please add an indication of how much RAM memory is required to run the job, what is the size of the output and how much time is required to complete the analysis;
- Concering the "Quick start" and "Examples" section:
   - the former is not "quick" since it requires 80TB of data, as well as a pre-processing of the data to solve a problem which is specific to a DESI data release;
   - The latter is only partially an "Examples" section since it relies on the outcome of the "Quick start" section.

  Hence, my suggestion is to move the *"Look into output files"* and *"Simple coadd showcase"* subsections into the *"Quick start"* section.  Also, bring the *"Examples for DESI members"* subsection to the top level (thus eliminating the *"Examples"* section).

- Fix the typo in the paper: syncronize -> synchronize;

Finally, I suggest to prepare and made available online the QSO catalog with the added LASTNIGHT column.  This is only about 25 MB in size and would simplify usage by a newcomer.  However this is a really minor point and is not mandatory for acceptance.
