### 2024-03-25 #6373 @gcalderone
Thanks @p-slash for the updates!
Please find my new comment in the same issue: https://github.com/p-slash/qsonic/issues/93
