### 2024-03-25 p-slash/qsonic#92 @corentinravoux
Thank you, you address all my comments here !
