### 2024-03-25 #6373 @corentinravoux
Thanks @p-slash, you answered all my comments. I have checked all the items of my checklist, and I recommend this package for publication.

I have a minor comment which is very technical and does not interfere at all with the review:

> > ```
> > * "a polynomial of ln 𝜆" Maybe I am wrong but i thought that for DESI, it is just "𝜆" ?
> > ```
>
> We are still fitting for ln lambda for DESI. There was a version of picca that performed the continuum fitting on lambda, but that was eventually removed before EDR analyses took off.

On this aspect, I was sure that it was in lambda in picca, but after double checking it is indeed in log_lambda. I do not understand why this is the case for DESI data which are linearly binned, did you test on QSOnic the difference in term of continuum fitting with different term in the polynomial term (lambda vs log lambda) ?
