### 2024-03-27 226414b
print nspec
