### 2024-03-27 801899d
fix: cont_params are valid
