+++ docs/source/quick_start.rst
@@ -1,4 +1,4 @@
-Quick Start
+Tutorial
 ===========
 
 Running qsonic-fit
@@ -54,7 +54,7 @@ Now we can pass this catalog into ``qsonic-fit`` script as the quasar catalog.
 
 .. code-block:: shell
 
-    srun -n 128 -c 2 qsonic-fit \
+    srun -n 16 -c 2 qsonic-fit \
     --input-dir ${EDR_DIRECTORY}/spectro/redux/fuji/healpix \
     --catalog QSO_cat_fuji_healpix_only_qso_targets_sv3_fix.fits \
     --keep-surveys sv3 \
@@ -64,7 +64,269 @@ Now we can pass this catalog into ``qsonic-fit`` script as the quasar catalog.
     --forest-w1 1040 --forest-w2 1200 \
     --skip-resomat
 
-Note, we passed ``--keep-surveys sv3`` argument to the script, and you still need to pass a valid output folder.
+Note, we passed ``--keep-surveys sv3`` argument to the script, and you still need to pass a valid output folder. This should read 
+
+Look into output files
+----------------------
+
+An empty notebook can be found in the GitHub repo under ``docs/nb/look_into_output_files.ipynb`` or downloaded :download:`here <../nb/look_into_output_files.ipynb>`.
+
+.. code:: python3
+
+    import fitsio
+    import numpy as np
+    import matplotlib.pyplot as plt
+
+    # Point to output delta folder
+    output_delta_folder = "Delta-co1"
+    fchi2 = fitsio.FITS(f"{output_delta_folder}/continuum_chi2_catalog.fits")[1]
+    fchi2
+
+Output:
+
+.. parsed-literal::
+
+    
+      file: Delta-co1/continuum_chi2_catalog.fits
+      extension: 1
+      type: BINARY_TBL
+      extname: CHI2_CAT
+      rows: 411359
+      column info:
+        TARGETID            i8  
+        Z                   f4  
+        HPXPIXEL            i8  
+        ARMS                S5  
+        MEANSNR             f4  array[3]
+        RSNR                f4  
+        MPI_RANK            i4  
+        CONT_valid          b1  
+        CONT_chi2           f4  
+        CONT_dof            i4  
+        CONT_x              f4  array[2]
+        CONT_xcov           f4  array[4]
+
+
+
+.. code:: python3
+
+    chi2_data = fchi2.read()
+    is_valid = chi2_data['CONT_valid']
+
+    chi2_v = chi2_data[is_valid]['CONT_chi2'] / chi2_data[is_valid]['CONT_dof']
+    
+    plt.hist(chi2_v, bins=100)
+    plt.axvline(1, c='k')
+    plt.xlabel(r"$\chi^2_\nu$")
+    plt.ylabel("Counts")
+    plt.yscale("log")
+    plt.show()
+
+
+
+.. image:: _static/chi2cat_hist.png
+
+
+Now let us investigate the ``attributes.fits`` file, which contains the mean continuum in ``CONT-i`` extensions, stacked fluxes in observed frame in ``STACKED_FLUX-i``, in rest-frame in ``STACKED_FLUX_RF-i`` extensions, and varlss-eta values in ``VAR_FUNC-i`` extenstions for all iterations.
+
+.. code:: python3
+
+    fattr = fitsio.FITS(f"{output_delta_folder}/attributes.fits")
+    fattr
+
+Output:
+
+.. parsed-literal::
+
+    
+      file: Delta-co1/attributes.fits
+      extnum hdutype         hduname[v]
+      0      IMAGE_HDU       
+      1      BINARY_TBL      CONT-1
[…]
