+++ docs/source/index.rst
@@ -32,7 +32,7 @@ See :doc:`/generated/modules` for details.
    :maxdepth: 2
 
    installation
-   quick_start
+   tutorial
    examples/desi_member_examples
    howitworks
    contributing
