+++ docs/source/paper.md
@@ -32,7 +32,7 @@ The density of the intervening gas is inferred from the transmitted flux fractio
 
 # Statement of need
 
-`QSOnic` is an MPI-parallelized, highly optimized Python package specifically built for the Dark Energy Spectroscopic Instrument (DESI), which will collect approximately almost a million Ly$\alpha$ quasar spectra in the near future [@desicollaborationDESIExperimentPart:2016]. `QSOnic` is designed to overcome the numerical challenge posed by this large quantity of spectra by MPI parallelization that can distribute the memory and workload across multiple compute nodes. Each parallel process is responsible for independent sky regions based on HEALPix [@healpix]. Reading and fitting the continuum are done in parallel. After all quasars are fit with a continuum, all processes syncronize to calculate the mean continuum and check for convergence.
+`QSOnic` is an MPI-parallelized, highly optimized Python package specifically built for the Dark Energy Spectroscopic Instrument (DESI), which will collect approximately almost a million Ly$\alpha$ quasar spectra in the near future [@desicollaborationDESIExperimentPart:2016]. `QSOnic` is designed to overcome the numerical challenge posed by this large quantity of spectra by MPI parallelization that can distribute the memory and workload across multiple compute nodes. Each parallel process is responsible for independent sky regions based on HEALPix [@healpix]. Reading the data and fitting the continuum are done in parallel. After all quasars are fit with a continuum, all processes synchronize to calculate the mean continuum and check for convergence.
 
 Furthermore, `QSOnic` provides an efficient API to read and manipulate DESI quasar spectra to allow implementation of other algorithms within the Ly$\alpha$ forest framework. It saves detailed intermediate data products including best-fit parameters for each quasar and the covariance matrix used in pipeline noise calibration correction ($\eta$ parameter in references [@bourbouxCompletedSDSSIVExtended:2020; @karacayliOptimal1DLy:2024]). These intermediate products are crucial to ascertain the precision of the fitted continua and the pipeline noise corrections.
 
