### 2024-03-29 a02a6ca
RAM, time and storage requirements
