### 2024-03-29 p-slash/qsonic#93 @p-slash
Thanks. I made the changes you asked for. I think they are reasonable.

>   In the "Quick start" section you correctly mention the size of the dataset (80 TB).  Please add an indication of how much RAM memory is required to run the job, what is the size of the output and how much time is required to complete the analysis;
> ...
>   Finally, I suggest to prepare and made available online the QSO catalog with the added LASTNIGHT column. This is only about 25 MB in size and would simplify usage by a newcomer. However this is a really minor point and is not mandatory for acceptance.

I added the following in the ``Workaround for EDR catalogs`` section: This catalog has 38,289 objects, of which ~11,000 are at redshifts z > 2.1 and have their Lya forest region within DESI’s wavelength coverage. You can find an unofficial copy of this catalog here (which is a Google Drive link).

Also the following: This minimal example should calculate quasar continua for about 11,000 spectra. It requires less than 1 GB of RAM and should finish in about a minute for 16 MPI tasks with two cores each. The output files require 320 MB storage. Reading the resolution matrix (by removing --skip-resomat option) multiplies the storage and RAM requirements by a factor of four and slows down reading/writing data. This slow down is neglible for this example, but are relevant for large samples.

>     * Concering the "Quick start" and "Examples" section:
>
>       * the former is not "quick" since it requires 80TB of data, as well as a pre-processing of the data to solve a problem which is specific to a DESI data release;
>       * The latter is only partially an "Examples" section since it relies on the outcome of the "Quick start" section.
>
>       Hence, my suggestion is to move the _"Look into output files"_ and _"Simple coadd showcase"_ subsections into the _"Quick start"_ section.  Also, bring the _"Examples for DESI members"_ subsection to the top level (thus eliminating the _"Examples"_ section).

The section is now titled `Tutorial`. The subsections are merged into it. DESI examples have their section.

>     * Fix the typo in the paper: syncronize -> synchronize;

Fixed.
