### 2024-03-29 #6373 @p-slash
Thanks. I addressed @gcalderone's comments under the issue.

@corentinravoux, re: ln lambda: Yes, DESI is linearly spaced, but this term is related to the quasar continuum model, so it does not need to be tied to the observational grid. If I remember correctly, ln lambda is the better of the two for the quasar continuum model theoretically, but I do not have a document to point you to. I had the comparison with picca, and the difference was small, but I did not code within qsonic. Furthermore, changing this model then propagates to the distortion matrix calculation. It complicates the entire analysis, so it may be that we simplified it to a single choice in the end.
