### 2024-04-01 4a398da
unit test for delta coadding
