### 2024-04-01 17ba0d1
write function for Delta class. ra, dec in degrees as properties of Delta class
