### 2024-04-01 5b94375
configure logger
