### 2024-04-01 5f8fc08
read all directories with a single pool
