### 2024-04-01 091348c
fix: write_one_hpx
