### 2024-04-02 16263f9
set deltas to 0 when ivar==0
