### 2024-04-02 924e1f7
set ivar==0 deltas to zero in Delta class
