### 2024-04-02 2fd8dff
safety against negative weights
