### 2024-04-02 8c15820
fix: additive var_lss correction should be done after continuum fitting
