### 2024-04-03 p-slash/qsonic#93 @gcalderone
Looks good to me, thanks!
I close this issue and recommend acceptance of the paper.
