### 2024-04-03 #6373 @gcalderone
Thanks @p-slash !
I closed the https://github.com/p-slash/qsonic/issues/93 issue, and recommend acceptance of the paper.
