### 2024-04-29 11abaa3
Editoral changes to paper.md
