### 2024-04-29 #6373 @ivastar
The DOI for the DESI Experiment Part I paper is https://doi.org/10.48550/arXiv.1611.00036

`picca` should just be cited with the du Max Bourboux paper as per the instructions on the repo:
https://doi.org/10.48550/arXiv.2007.08995

Some editorial comments in https://github.com/p-slash/qsonic/pull/96
