+++ docs/source/paper.bib
@@ -124,6 +124,7 @@ archivePrefix = {arXiv},
     month = oct,
     eid = {arXiv:1611.00036},
     pages = {arXiv:1611.00036},
+    doi = {10.48550/arXiv.1611.00036},
     archivePrefix = {arXiv},
     eprint = {1611.00036},
     primaryClass = {astro-ph.IM},
