### 2024-05-01 #6373 @p-slash
> `picca` should just be cited with the du Max Bourboux paper as per the instructions on the repo:

@ivastar could you clarify this comment? Do you want me to remove citations or add this citation instead of a link to the repo? Note that this paper is already cited.
