### 2024-05-01 01ef954
reading exposures before continuum option
