### 2024-05-01 1c35928
set forestflux=0 in mask.py
