### 2024-05-01 c684e49
get_median_outlier_mask function implemented in mathtools.py
