### 2024-05-01 90ef7e0
outliers are masked in set_forest_region function in spectrum.py
