+++ py/qsonic/spectrum.py
@@ -2,7 +2,8 @@ import argparse
 
 import numpy as np
 
-from qsonic.mathtools import _zero_function, _one_function, get_smooth_ivar
+from qsonic.mathtools import (
+    _zero_function, _one_function, get_smooth_ivar, get_median_outlier_mask)
 
 
 def add_wave_region_parser(parser=None):
@@ -282,8 +283,11 @@ class Spectrum():
         self.cont_params['x'][0] /= cont_params_weight
 
     def set_forest_region(self, w1, w2, lya1, lya2):
-        """ Sets slices for the forest region. Also calculates the mean SNR in
-        the forest and an initial guess for the continuum amplitude.
+        """ Sets slices for the forest region. Masks outliers in each arm
+        separately based on median statistics
+        (see :func:`qsonic.mathtools.get_median_outlier_mask`). Also calculates
+        the mean SNR in the forest and an initial guess for the continuum
+        amplitude.
 
         Arguments
         ---------
@@ -304,7 +308,6 @@ class Spectrum():
 
             self._f1[arm], self._f2[arm] = ii1, ii2
 
-            # Does this create a view or copy array?
             # See https://numpy.org/doc/stable/user/basics.copies.html
             # Slicing creates views, not copies. Bases of views are not removed
             # from memory!
@@ -313,6 +316,11 @@ class Spectrum():
             self._forestivar[arm] = self.ivar[arm][ii1:ii2].copy()
             if self.reso:
                 self._forestreso[arm] = self.reso[arm][:, ii1:ii2].copy()
+
+            w = get_median_outlier_mask(
+                self._forestflux[arm], self._forestivar[arm])
+            self._forestflux[arm][w] = 0
+            self._forestivar[arm][w] = 0
 
         self._forestivar_sm = self._forestivar
         self._forestweight = self._forestivar
