### 2024-05-01 2b53a73
test outlier masking in test_spectrum.py
