+++ tests/test_spectrum.py
@@ -57,6 +57,11 @@ class TestSpectrum(object):
         assert ('R' not in spec.forestflux.keys())
         assert (not spec.forestreso)
 
+        spec.flux['B'][10:15] = 1e2
+        spec.ivar['B'][10:15] = 0
+        spec.set_forest_region(3600., 6000., 1050., 1180.)
+        npt.assert_equal(spec.forestflux['B'][10:15], 0)
+
     def test_drop_short_arms(self, setup_data):
         cat_by_survey, _, data = setup_data(1)
         spectra_list = qsonic.spectrum.generate_spectra_list_from_data(
