### 2024-05-01 0b8c808
update packages. remove tests for python3.8 and include test for python3.11
