### 2024-05-01 12e1079
fix testing outlier masking in test_spectrum.py
