### 2024-05-02 d4ad75b
remove unused bib entry for picca in paper.bib
