+++ docs/source/paper.bib
@@ -106,15 +106,6 @@ archivePrefix = {arXiv},
 }
 
 
-@misc{picca,
-  author = {{du Mas des Bourboux}, H{\'e}lion and {P{\'e}rez-R{\`a}fols}, Ignasi and {Walther}, Michael},
-  title = {picca: Package for Igm Cosmological-Correlations Analyses},
-  year = {2024},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/igmhub/picca}
-}
-
 @ARTICLE{desicollaborationDESIExperimentPart:2016,
     author = {{DESI Collaboration} and {Aghamousa}, Amir and {Aguilar}, Jessica and {Ahlen}, Steve and {Alam}, Shadab and {Allen}, Lori E. and {Allende Prieto}, Carlos and {Annis}, James and {Bailey}, Stephen and {Balland}, Christophe and {Ballester}, Otger and {Baltay}, Charles and {Beaufore}, Lucas and {Bebek}, Chris and {Beers}, Timothy C. and {Bell}, Eric F. and {Bernal}, Jos{\'e} Luis and {Besuner}, Robert and {Beutler}, Florian and {Blake}, Chris and {Bleuler}, Hannes and {Blomqvist}, Michael and {Blum}, Robert and {Bolton}, Adam S. and {Briceno}, Cesar and {Brooks}, David and {Brownstein}, Joel R. and {Buckley-Geer}, Elizabeth and {Burden}, Angela and {Burtin}, Etienne and {Busca}, Nicolas G. and {Cahn}, Robert N. and {Cai}, Yan-Chuan and {Cardiel-Sas}, Laia and {Carlberg}, Raymond G. and {Carton}, Pierre-Henri and {Casas}, Ricard and {Castander}, Francisco J. and {Cervantes-Cota}, Jorge L. and {Claybaugh}, Todd M. and {Close}, Madeline and {Coker}, Carl T. and {Cole}, Shaun and {Comparat}, Johan and {Cooper}, Andrew P. and {Cousinou}, M. -C. and {Crocce}, Martin and {Cuby}, Jean-Gabriel and {Cunningham}, Daniel P. and {Davis}, Tamara M. and {Dawson}, Kyle S. and {de la Macorra}, Axel and {De Vicente}, Juan and {Delubac}, Timoth{\'e}e and {Derwent}, Mark and {Dey}, Arjun and {Dhungana}, Govinda and {Ding}, Zhejie and {Doel}, Peter and {Duan}, Yutong T. and {Ealet}, Anne and {Edelstein}, Jerry and {Eftekharzadeh}, Sarah and {Eisenstein}, Daniel J. and {Elliott}, Ann and {Escoffier}, St{\'e}phanie and {Evatt}, Matthew and {Fagrelius}, Parker and {Fan}, Xiaohui and {Fanning}, Kevin and {Farahi}, Arya and {Farihi}, Jay and {Favole}, Ginevra and {Feng}, Yu and {Fernandez}, Enrique and {Findlay}, Joseph R. and {Finkbeiner}, Douglas P. and {Fitzpatrick}, Michael J. and {Flaugher}, Brenna and {Flender}, Samuel and {Font-Ribera}, Andreu and {Forero-Romero}, Jaime E. and {Fosalba}, Pablo and {Frenk}, Carlos S. and {Fumagalli}, Michele and {Gaensicke}, Boris T. and {Gallo}, Giuseppe and {Garcia-Bellido}, Juan and {Gaztanaga}, Enrique and {Pietro Gentile Fusillo}, Nicola and {Gerard}, Terry and {Gershkovich}, Irena and {Giannantonio}, 
[…]
