### 2024-05-02 #6373 @p-slash
> > @ivastar could you clarify this comment? Do you want me to remove Karaçaylı et al. (2024) citations or add this citation instead of a link to the repo? Note that this paper is already cited.
>
> I think use the paper citation instead of the link to the repo. Just trying to resolve the "missing DOI" issues.

Ok. The problem was an unused entry in the bib file, not the link to the repo. No more missing DOIs now
