### 2024-05-02 512b552
github auto tests ignore docs/** paths
