+++ .github/workflows/testing.yml
@@ -1,6 +1,9 @@
 name: Tests
 
-on: [push]
+on:
+  push:
+    paths-ignore:
+      - 'docs/**'
 
 jobs:
   build:
