### 2024-05-02 0ee38e1
[skip ci] Bump version: 0.8.3 → 1.0.0
