+++ py/qsonic/__init__.py
@@ -1,6 +1,6 @@
 __author__ = 'Naim Goksel Karacayli'
 __email__ = 'ngokselk@gmail.com'
-__version__ = '0.8.3'
+__version__ = '1.0.0'
 
 
 class QsonicException(Exception):
