### 2024-05-02 cba5b43
update exposures description in additional notes in howitworks.rst
