### 2024-05-04 90479a3
BaseContinuumModel implemented
