### 2024-05-04 d912553
PiccaContinuumModel moves fit_continuum, get_continuum_model, _continuum_costfn and minimizer functions.
