+++ py/qsonic/continuum_models/picca_continuum_model.py
@@ -1,0 +1,246 @@
+"""Picca continuum model."""
+import numpy as np
+from iminuit import Minuit
+from scipy.optimize import minimize
+
+from qsonic import QsonicException
+from qsonic.mathtools import mypoly1d
+from qsonic.continuum_models.base_continuum_model import BaseContinuumModel
+
+
+class PiccaContinuumModel(BaseContinuumModel):
+    """Picca continuum model class.
+
+    Parameters
+    ----------
+    meancont_interp: FastCubic1DInterp
+        Fast cubic spline object for the mean continuum.
+    meanflux_interp: FastLinear1DInterp
+        Interpolator for mean flux. If fiducial is not set, this equals to 1.
+    varlss_interp: FastLinear1DInterp or FastCubic1DInterp
+        Cubic spline for var_lss if fitting. Linear if from file.
+    eta_interp: FastCubic1DInterp
+        Interpolator for eta. Returns one if fiducial var_lss is set.
+    cont_order: int
+        Order of continuum polynomial from ``args.cont_order``.
+    rfwave0: float
+        First rest-frame wavelength center for the mean continuum.
+    denom: float
+        Denominator for the slope term in the continuum model.
+    minimizer: str
+        ``iminuit`` or ``l_bfgs_b`` to select the minimizer function.
+
+    Attributes
+    ----------
+    minimizer: function
+        Function that points to one of the minimizer options.
+    """
+
+    def __init__(
+            self, meancont_interp, meanflux_interp, varlss_interp,
+            eta_interp, cont_order, rfwave0, denom, minimizer
+    ):
+        self.meancont_interp = meancont_interp
+        self.meanflux_interp = meanflux_interp
+        self.varlss_interp = varlss_interp
+        self.eta_interp = eta_interp
+        self.cont_order = cont_order
+        self.rfwave0 = rfwave0
+        self.denom = denom
+
+        if minimizer == "iminuit":
+            self.minimizer = self._iminuit_minimizer
+        elif minimizer == "l_bfgs_b":
+            self.minimizer = self._scipy_l_bfgs_b_minimizer
+        else:
+            raise QsonicException(
+                "Undefined minimizer. Developer forgot to implement.")
+
+    def _continuum_costfn(self, x, wave, flux, ivar_sm, z_qso):
+        """Cost function to minimize for each quasar.
+
+        This is a modified chi2 where amplitude is also part of minimization.
+        Cost of each arm is simply added to the total cost.
+
+        Arguments
+        ---------
+        x: :external+numpy:py:class:`ndarray <numpy.ndarray>`
+            Polynomial coefficients for quasar diversity.
+        wave: dict(:external+numpy:py:class:`ndarray <numpy.ndarray>`)
+            Observed-frame wavelengths.
+        flux: dict(:external+numpy:py:class:`ndarray <numpy.ndarray>`)
+            Flux.
+        ivar_sm: dict(:external+numpy:py:class:`ndarray <numpy.ndarray>`)
+            Smooth inverse variance.
+        z_qso: float
+            Quasar redshift.
+
+        Returns
+        ---------
+        cost: float
+            Cost (modified chi2) for a given ``x``.
+        """
+        cost = 0
+
+        for arm, wave_arm in wave.items():
+            cont_est = self.get_continuum_model(x, wave_arm / (1 + z_qso))
+            # no_neg = np.sum(cont_est<0)
+            # penalty = wave_arm.size * no_neg**2
+
+            cont_est *= self.meanflux_interp(wave_arm)
+
+            var_lss = self.varlss_interp(wave_arm) * cont_est**2
+            eta = self.eta_interp(wave_arm)
[…]
