### 2024-05-04 31e52ce
fit_continuum, get_continuum_model, _continuum_costfn and minimizer functions are moved to picca_continuum_model.py
