+++ py/qsonic/picca_continuum.py
@@ -5,8 +5,7 @@ import logging
 import numpy as np
 from numba import njit
 import fitsio
-from iminuit import Minuit
-from scipy.optimize import minimize, curve_fit
+from scipy.optimize import curve_fit
 from scipy.interpolate import UnivariateSpline, CubicSpline
 from scipy.special import legendre
 
@@ -16,10 +15,12 @@ from qsonic import QsonicException
 from qsonic.spectrum import valid_spectra
 from qsonic.mpi_utils import mpi_fnc_bcast, MPISaver
 from qsonic.mathtools import (
-    mypoly1d, block_covariance_of_square,
+    block_covariance_of_square,
     FastLinear1DInterp, FastCubic1DInterp,
     SubsampleCov
 )
+
+from qsonic.continuum_models.picca_continuum_model import PiccaContinuumModel
 
 
 def add_picca_continuum_parser(parser=None):
@@ -108,8 +109,6 @@ class PiccaContinuumFitter():
         Denominator for the slope term in the continuum model.
     meancont_interp: FastCubic1DInterp
         Fast cubic spline object for the mean continuum.
-    minimizer: function
-        Function that points to one of the minimizer options.
     comm: MPI.COMM_WORLD
         MPI comm object to reduce, broadcast etc.
     mpi_rank: int
@@ -136,6 +135,8 @@ class PiccaContinuumFitter():
         Normalizes observed flux to be 1 if True.
     eta_calib_ivar: bool
         Calibrate IVAR with eta estimates.
+    model: Derived(BaseContinuumModel)
+        Continuum model to use
     """
 
     def _get_fiducial_interp(self, fname, col2read):
@@ -204,14 +205,6 @@ class PiccaContinuumFitter():
             self.rfwave[0], self.dwrf, np.ones(self.nbins),
             ep=np.zeros(self.nbins))
 
-        if args.minimizer == "iminuit":
-            self.minimizer = self._iminuit_minimizer
-        elif args.minimizer == "l_bfgs_b":
-            self.minimizer = self._scipy_l_bfgs_b_minimizer
-        else:
-            raise QsonicException(
-                "Undefined minimizer. Developer forgot to implement.")
-
         self.comm = MPI.COMM_WORLD
         self.mpi_rank = self.comm.Get_rank()
 
@@ -247,176 +240,10 @@ class PiccaContinuumFitter():
         self.normalize_stacked_flux = args.normalize_stacked_flux
         self.eta_calib_ivar = args.eta_calib_ivar
 
-    def _continuum_costfn(self, x, wave, flux, ivar_sm, z_qso):
-        """ Cost function to minimize for each quasar.
-
-        This is a modified chi2 where amplitude is also part of minimization.
-        Cost of each arm is simply added to the total cost.
-
-        Arguments
-        ---------
-        x: :external+numpy:py:class:`ndarray <numpy.ndarray>`
-            Polynomial coefficients for quasar diversity.
-        wave: dict(:external+numpy:py:class:`ndarray <numpy.ndarray>`)
-            Observed-frame wavelengths.
-        flux: dict(:external+numpy:py:class:`ndarray <numpy.ndarray>`)
-            Flux.
-        ivar_sm: dict(:external+numpy:py:class:`ndarray <numpy.ndarray>`)
-            Smooth inverse variance.
-        z_qso: float
-            Quasar redshift.
-
-        Returns
-        ---------
-        cost: float
-            Cost (modified chi2) for a given ``x``.
-        """
-        cost = 0
-
-        for arm, wave_arm in wave.items():
-            cont_est = self.get_continuum_model(x, wave_arm / (1 + z_qso))
-            # no_neg = np.sum(cont_est<0)
-            # penalty = wave_arm.size * no_neg**2
-
[…]
