+++ docs/source/generated/continuum_models.rst
@@ -1,0 +1,21 @@
+qsonic.continuum_models
+=======================
+
+qsonic.continuum_models.base_continuum_model
+----------------------------------------------
+
+
+.. automodule:: qsonic.continuum_models.base_continuum_model
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+qsonic.continuum_models.picca_continuum_model
+----------------------------------------------
+
+
+.. automodule:: qsonic.continuum_models.picca_continuum_model
+   :members:
+   :undoc-members:
+   :show-inheritance:
