### 2024-05-05 9cffd1b
move true continuum analysis to true_continuum_model.py module
