+++ py/qsonic/picca_continuum.py
@@ -20,8 +20,6 @@ from qsonic.mathtools import (
     SubsampleCov
 )
 
-from qsonic.continuum_models.picca_continuum_model import PiccaContinuumModel
-
 
 def add_picca_continuum_parser(parser=None):
     """ Adds PiccaContinuumFitter related arguments to parser. These
@@ -46,8 +44,13 @@ def add_picca_continuum_parser(parser=None):
         "--num-iterations", type=int, default=10,
         help="Number of iterations for continuum fitting.")
     cont_group.add_argument(
+        "--continuum-model", default="picca", choices=["picca", "true"],
+        help="Continuum model. *picca* fits the continuum in the forest "
+        "region. *true* is the true continuum analysis for mock analysis."
+    )
+    cont_group.add_argument(
         "--true-continuum", action="store_true",
-        help="True continuum analysis deltas if mock analysis")
+        help="Alternative argument for true continuum analysis.")
     cont_group.add_argument(
         "--fiducial-meanflux", help="Fiducial mean flux FITS file.")
     cont_group.add_argument(
@@ -194,6 +197,7 @@ class PiccaContinuumFitter():
         # We first decide how many bins will approximately satisfy
         # rest-frame wavelength spacing. Then we create wavelength edges, and
         # transform these edges into centers
+        self.args = args
         self.nbins = int(round(
             (args.forest_w2 - args.forest_w1) / args.rfdwave))
         edges, self.dwrf = np.linspace(
@@ -240,10 +244,23 @@ class PiccaContinuumFitter():
         self.normalize_stacked_flux = args.normalize_stacked_flux
         self.eta_calib_ivar = args.eta_calib_ivar
 
-        self.model = PiccaContinuumModel(
-            self.meancont_interp, self.meanflux_interp, self.varlss_interp,
-            self.eta_interp, self.cont_order, self.rfwave[0], self._denom,
-            args.minimizer)
+        if args.continuum_model == "picca":
+            from qsonic.continuum_models.picca_continuum_model import \
+                PiccaContinuumModel
+
+            self.model = PiccaContinuumModel(
+                self.meancont_interp, self.meanflux_interp, self.varlss_interp,
+                self.eta_interp, self.cont_order, self.rfwave[0], self._denom,
+                args.minimizer)
+
+        elif args.continuum_model == "true":
+            from qsonic.continuum_models.true_continuum_model import \
+                TrueContinuumModel
+
+            self.model = TrueContinuumModel(
+                self.meanflux_interp, self.varlss_interp)
+            self.cont_order = 0
+            self.niterations = 1
 
     def fit_continua(self, spectra_list):
         """ Fits all continua for a list of Spectrum objects.
@@ -547,72 +564,24 @@ class PiccaContinuumFitter():
         self._eta_calibate_ivar(spectra_list)
 
         self.save(fattr)
-        if self.varlss_fitter:
-            self.varlss_fitter.write(fattr)
+        if self.varlss_fitter is None:
+            self.fit_final_varlss(spectra_list)
+            self.save(fattr, '-fit')
+
+        self.varlss_fitter.write(fattr)
         fattr.close()
         logging.info("All continua are fit.")
 
-    def true_continuum(self, spectra_list, args):
-        """True continuum reduction. Uses fiducials for mean flux and varlss
-        interpolation. Continuum is interpolated using a cubic spline.
-
-        .. warning::
-
-            This function would work even if you did not pass any fiducial.
-
-        Arguments
-        ---------
-        spectra_list: list(Spectrum)
-            Spectrum objects.
-        args: argparse.Namespace
[…]
