+++ py/qsonic/scripts/qsonic_fit.py
@@ -60,6 +60,9 @@ def get_parser(add_help=True):
 def args_logic_fnc_qsonic_fit(args):
     args.arms = list(set(args.arms))
 
+    is_true_continuum = args.continuum_model == "true" or args.true_continuum
+    args.true_continuum = is_true_continuum
+    args.continuum_model = "true"
     condition_msg = [
         (args.true_continuum and not args.mock_analysis,
             "True continuum is only applicable to mock analysis."),
@@ -260,16 +263,12 @@ def mpi_continuum_fitting(spectra_list, args, comm, mpi_rank):
     start_time = time.time()
     qcfit = PiccaContinuumFitter(args)
 
-    if not args.true_continuum:
-        # Fit continua
-        # Stack all spectra in each process
-        # Broadcast and recalculate global functions
-        # Iterate
-        logging.info("Fitting continuum.")
-        qcfit.iterate(spectra_list)
-    else:
-        logging.info("True continuum.")
-        qcfit.true_continuum(spectra_list, args)
+    # Fit continua
+    # Stack all spectra in each process
+    # Broadcast and recalculate global functions
+    # Iterate
+    logging.info("Fitting continuum.")
+    qcfit.iterate(spectra_list)
 
     valid_spectra = list(qsonic.spectrum.valid_spectra(spectra_list))
 
